optimize modal dialog
