--- src/components/ModalContainer/ModalContainer.scss
@@ -1,21 +1,0 @@
-.modal-container {
-  padding: 15px;
-
-  &-header {
-    text-align: center;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 20px;
-
-    & > .title {
-      font-size: 22px;
-      line-height: 24px;
-      font-weight: 700;
-    }
-
-    & > .icon {
-      cursor: pointer;
-    }
-  }
-}
